Update np.float and r-value propagation
1. Update "np.finfo(np.float)" to "np.finfo(float)" to be compatible with future NumPy versions

2. Propagate correlation coefficients (spm.r) from SPM objects to SPM inference objects (spmi.r)

+++ spm1d/rft1d/geom.py
@@ -5,14 +5,15 @@ This module contains functions for computing various geomtric characteristics
 of 1D fields and upcrossings.
 '''
 
-# Copyright (C) 2016  Todd Pataky
+# Copyright (C) 2022  Todd Pataky
 
 
 from math import log
 import numpy as np
 
 
-eps         = np.finfo(np.float).eps
+# eps         = np.finfo(np.float).eps
+eps         = np.finfo(float).eps
