+++ spm1d/rft1d/prob.py
@@ -5,7 +5,7 @@ The core RFT computations are conducted inside **prob.rft**, and the
 **RFTCalculator** class serves as a high-level interface to **prob.rft**
 '''
 
-# Copyright (C) 2016  Todd Pataky
+# Copyright (C) 2022  Todd Pataky
 
 
 
@@ -20,7 +20,8 @@ FOUR_LOG2   = 4*log(2)
 SQRT_4LOG2  = sqrt(4*log(2))
 SQRT_2      = sqrt(2)
 TWO_PI      = 2*pi
-eps         = np.finfo(np.float).eps
+# eps         = np.finfo(np.float).eps
+eps         = np.finfo(float).eps
